debug filterShapes and match network to citation
If widenFactor is greater than 1, the input shape of the first residual block had the wrong number of input channels.  Similarly the same filterShape cannot be used for both conv1 and conv2 when the number of channels is being increased.

Added identity connections, dropout between convolutions, and the use of preactivation in the 1x1 conv2D shortcuts.  

The paper also uses a weight decay of 5e-4.  I can add a function for this but I have not seen any implementations of L2 decay in any of the models so I wondered if there was some plan to add it somewhere else in the API.