FastStyleTransfer: use `roundf` rather than `round` (#304)
This is needed to build on Windows.  Without this, the build would fail
as the expression would fail to typecheck.  Windows does not provide the
`round` macro, and requires that we explicitly use the typed macro for
the floating point operation (e.g. `roundf`, `roundl`, `roundll` instead
of `round`).

With this change it is possible to build the models on Windows.