TransformerUI: add initial macOS UI
This is modelled after the Windows UI, and looks very similar.  This
actually helps demonstrate the point that the usage of the model is
extremely similar.  This sets an initial starting point for
improvements.

The differences from the Windows implementation include the fact that
`\n` does not need to be replaced with `\r\n`, the scrollview for the
textview needs to be constructed manually, and the coordinate system
uses a different origin.