build: remove TransformerDemo from installation
We do not currently install any of the demos, remove the installation
for this particular demo.  We could introduce a new option to install
the demo programs, but given that they are demo programs, they are not
useful to install for persistence but more as a guide for getting
started.