Remove device placement
For the CIFAR10 and MNIST examples, the dataset is already loaded on the device, so these lines (which also fail with x10) will provoke a copy to the CPU then back on the accelerator, which will take unnecessary time.

For other datasets, there is probably no change doing it this way. This was the only part of `Tensorflow` used, so it should be okay to remove the import too.