[WIP]: Exploring removing `<Float>` everywhere.
This change explores stuffing some extra typealiases in a protocol to obviate
the need to specify `Foo<Float>` everywhere. Unfortunately, it appears that
the typealiases don't override the imports when they're defined in another
module.

Disadvantages of this approach: it might make it more confusing for someone
who is unfamiliar with this pattern to understand what is specifying the
scalar parameter.

Finally, in certain usages, typealiases even within the same module did not
work, and had to use `Self.$TYPEALIAS` to get it to compile, even in the
same module. Also, although I didn't play around with it much, the `Tensor`
typealias also caused a number of problems in the typechecker.